requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.1

langchain>=0.2.11
langchain-core>=0.2.30
langchain-groq>=0.1.6

langgraph>=0.2.35

yfinance>=0.2.40
pandas>=2.2.2
numpy>=2.0.1

requests>=2.32.3
feedparser>=6.0.11

diskcache>=5.6.3